Hardware: Cast GPE EnableMask before storing
In function AcpiHwLowSetGpe(), cast EnableMask to UINT8 before
storing. The mask was read from a 32 bit register but is an 8 bit
value. Fixes Visual Studio compiler warning.

+++ source/components/hardware/hwgpe.c
@@ -239,7 +239,7 @@ AcpiHwLowSetGpe (
     Status = AcpiHwWrite (EnableMask, &GpeRegisterInfo->EnableAddress);
     if (ACPI_SUCCESS (Status) && (Action & ACPI_GPE_SAVE_MASK))
     {
-        GpeRegisterInfo->EnableMask = EnableMask;
+        GpeRegisterInfo->EnableMask = (UINT8) EnableMask;
     }
     return (Status);
 }
